Update README.md
This is a project i did in university in collaboration with @kristianperriu. Providing the lyrics of a song this model will predict what genre it is between rap, pop, rock and country. Here we provided 3 notebooks; generating the dataset, processing the dataset and a notebook containing the model. 
Find attached also a power-point presentation and a detailed report.